// src/models/winner.ts
export interface IJackpotWinner {
    "amount_cents": number;
    "jackpot_name": string;
    "jackpot_level_name": string;
    "approved_at": string;
    "created_at": string;
    "jackpot_level_external_code": string;
    "jackpot_external_code": string;
    "user_name": string;
    "jackpot_original_currency": string;
    "jackpot_original_amount_cents": number;
}